+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-### v0.0.1
+### v0.0.3-alpha
+- Feature `Config{}`
+- Feature `Pop()` to retrieve and remove a key-value
+- Feature implement id-disk storage
+
+### v0.0.2-alpha
+- Feature `expiration`
+
+### v0.0.1-alpha
 - Feature `Set()` to store key-value
 - Feature `Get()` to retrieve key-value